requirements.txt
torch>=1.10.0
transformers>=4.25.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.2
flask>=2.3.0
numpy>=1.21.0
requests>=2.28.0
tqdm>=4.64.0